Fix shape2str function to use size_t instead of unsigned long for vector type

+++ include/tensor/xtensor_lib.h
@@ -34,7 +34,7 @@ typedef xt::xarray<double> double_tensor;
 
 
 
-string shape2str(xt::svector<unsigned long> vec);
+string shape2str(xt::svector<size_t> vec);
 int positive_index(int idx, int size);
 xt::xarray<double> outer_stack(xt::xarray<double> X, xt::xarray<double>  Y);
 xt::xarray<double> diag_stack(xt::xarray<double> X);
